analyzers/TenableSC/requirements.txt
cortexutils
pytenable